debug edit n_qp to hash mode

+++ src/ray/object_manager/object_manager_rdma.h
@@ -141,7 +141,7 @@ public:
   void InsertObjectInQueue(std::vector<ObjectRdmaInfo> &object_rdma_info);
 
   void RunRdmaService(int64_t index);
-  void FetchObjectFromRemotePlasmaThreads(ObjectRdmaInfo &object_rdma_info, int t_index);
+  void FetchObjectFromRemotePlasmaThreads(ObjectRdmaInfo &object_rdma_info, int64_t t_index);
   int PollCompletionThreads(struct pingpong_context *ctx, const plasma::Allocation &allocation, const ray::ObjectInfo &object_info, const std::pair<const plasma::LocalObject *, plasma::flatbuf::PlasmaError>& pair, int64_t start_time, int64_t te_fetch_object_rdma_space, int64_t te_fetch_object_post_send);
